Updated the product cards with scroll conditonals that will turn arrow on and off based on location.
